ENH: Added LongLookupTask
ENH: Added LongLookupTask class, changed all json to yml, changed get_default_params appropriately, moved helper functions to separate file, added docstring for get_default_params.
