--- LongLookupTables/LongLookupTables/default_params.json
@@ -1,8 +1,0 @@
-{
-    "task_defaults": {
-        "batch_size": 1,
-        "k": 10,
-        "max_len": 10,
-        "patience": 15
-    }
-}
