--- LongLookupTables/NoisyLongLookupTablesMulti/default_params.json
@@ -1,8 +1,0 @@
-{
-    "task_defaults": {
-        "batch_size": 64,
-        "k": 3,
-        "max_len": 30,
-        "patience": 7
-    }
-}
